Signed return values where necessary

+++ bridge.h
@@ -3,10 +3,10 @@
 
 uint32_t bridge_releaseProcessor(void);
 uint32_t bridge_setProcessPriority(uint8_t pid, uint8_t priority);
-uint8_t bridge_getProcessPriority(uint8_t pid);
+int16_t bridge_getProcessPriority(uint8_t pid);
 
 void *bridge_acquireMemoryBlock(void);
-uint32_t bridge_releaseMemoryBlock(void *blk);
+int8_t bridge_releaseMemoryBlock(void *blk);
 uint8_t bridge_isOutOfMemory(void);
 
 #endif
